current working state
not sure if the cve searches are working correctly, but we got the fallback url to populate with a legitimate search... need to continue testing to be sure.

will need to take a closer look at the cve response parsing.

everything else hinges on being able to pull and parse valid cve information.

so to the testing grouds